# docs/09-api/openapi.yaml
openapi: 3.1.0
info:
  title: Lumine API
  description: Institutional AI-native quantitative hedge fund API
  version: 0.1.0
paths:
  /api/v1/portfolio/summary:
    get:
      tags:
      - portfolio
      summary: Get Portfolio Summary
      description: Return the current portfolio summary.
      operationId: get_portfolio_summary_api_v1_portfolio_summary_get
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio:
    get:
      tags:
      - portfolio
      summary: Get Portfolio
      description: Return the single (default) portfolio (B-06 portfolio CRUD).
      operationId: get_portfolio_api_v1_portfolio_get
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - portfolio
      summary: Create Portfolio
      description: Create a portfolio — single-portfolio v1 returns the default book.
      operationId: create_portfolio_api_v1_portfolio_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/{portfolio_id}:
    put:
      tags:
      - portfolio
      summary: Update Portfolio
      description: Rename/update the portfolio (demo echo; single-portfolio v1).
      operationId: update_portfolio_api_v1_portfolio__portfolio_id__put
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - portfolio
      summary: Delete Portfolio
      description: Delete a portfolio — v1 refuses the only book (safe default).
      operationId: delete_portfolio_api_v1_portfolio__portfolio_id__delete
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/{portfolio_id}/equity:
    get:
      tags:
      - portfolio
      summary: Get Equity Curve
      description: 'Equity curve (B-06) — deterministic demo series until storage
        wiring.


        Points: nav/equity with a mild upward drift and a drawdown dip at the

        2/3 mark, 1h spacing, newest first.'
      operationId: get_equity_curve_api_v1_portfolio__portfolio_id__equity_get
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_EquityPoint_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/{portfolio_id}/orders:
    delete:
      tags:
      - portfolio
      summary: Cancel All Orders
      description: Cancel every open order in the portfolio (B-06 cancel-all).
      operationId: cancel_all_orders_api_v1_portfolio__portfolio_id__orders_delete
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelAllResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/{portfolio_id}/simulate:
    post:
      tags:
      - portfolio
      summary: Simulate Trade
      description: Project portfolio impact of a hypothetical trade (what-if, no execution).
      operationId: simulate_trade_api_v1_portfolio__portfolio_id__simulate_post
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateTradeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateTradeResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/positions:
    get:
      tags:
      - portfolio
      summary: List Positions
      description: List open positions.
      operationId: list_positions_api_v1_portfolio_positions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_Position_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/positions/{position_id}:
    get:
      tags:
      - portfolio
      summary: Get Position
      description: Return a single open position.
      operationId: get_position_api_v1_portfolio_positions__position_id__get
      parameters:
      - name: position_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Position Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/portfolio/exposure:
    get:
      tags:
      - portfolio
      summary: Get Exposure
      description: Return exposure breakdown.
      operationId: get_exposure_api_v1_portfolio_exposure_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_ExposureSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CancelAllResult:
      properties:
        cancelled:
          type: integer
          title: Cancelled
        portfolio_id:
          type: string
          title: Portfolio Id
      type: object
      required:
      - cancelled
      - portfolio_id
      title: CancelAllResult
      description: Cancel-all outcome.
...
    CreatePortfolioRequest:
      properties:
        name:
          type: string
          title: Name
          default: default
        currency:
          type: string
          title: Currency
          default: USD
      type: object
      title: CreatePortfolioRequest
      description: 'Create a portfolio (single-portfolio v1: returns the default).'
...
    EquityPoint:
      properties:
        ts:
          type: string
          format: date-time
          title: Ts
        nav:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Nav
        equity:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Equity
        drawdown:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Drawdown
      type: object
      required:
      - ts
      - nav
      - equity
      - drawdown
      title: EquityPoint
      description: One point on the portfolio equity curve.
    ExposureSummary:
      properties:
        symbol:
          type: string
          title: Symbol
        notional:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Notional
        pct_of_nav:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Pct Of Nav
        correlated_bucket:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlated Bucket
      type: object
      required:
      - symbol
      - notional
      - pct_of_nav
      title: ExposureSummary
      description: Exposure breakdown per symbol/bucket.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedList_EquityPoint_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EquityPoint'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[EquityPoint]
    PaginatedList_ExposureSummary_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ExposureSummary'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[ExposureSummary]
...
    PaginatedList_Position_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Position'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[Position]
...
    PortfolioSummary:
      properties:
        portfolio_id:
          type: string
          title: Portfolio Id
        nav:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Nav
        cash:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Cash
        margin_used:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Margin Used
        open_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Open Pnl
        closed_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Closed Pnl
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - portfolio_id
      - nav
      - cash
      - margin_used
      - open_pnl
      - closed_pnl
      - timestamp
      title: PortfolioSummary
      description: High-level portfolio snapshot.
    Position:
      properties:
        position_id:
          type: string
          format: uuid
          title: Position Id
        portfolio_id:
          type: string
          title: Portfolio Id
        symbol:
          type: string
          title: Symbol
        direction:
          type: string
          enum:
          - long
          - short
          title: Direction
        volume:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Volume
        entry_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Entry Price
        current_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Current Price
        stop_loss:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Loss
        take_profit:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Take Profit
        unrealized_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Unrealized Pnl
        opened_at:
          type: string
          format: date-time
          title: Opened At
      type: object
      required:
      - position_id
      - portfolio_id
      - symbol
      - direction
      - volume
      - entry_price
      - unrealized_pnl
      - opened_at
      title: Position
      description: Open position view.
...
    SimulateTradeRequest:
      properties:
        symbol:
          type: string
          title: Symbol
        side:
          type: string
          enum:
          - buy
          - sell
          title: Side
        volume:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Volume
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price
      type: object
      required:
      - symbol
      - side
      - volume
      - price
      title: SimulateTradeRequest
      description: Payload for a what-if trade simulation.
    SimulateTradeResult:
      properties:
        projected_nav:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Projected Nav
        margin_required:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Margin Required
        pnl_change:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Pnl Change
      type: object
      required:
      - projected_nav
      - margin_required
      - pnl_change
      title: SimulateTradeResult
      description: Projected portfolio impact of a hypothetical trade.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...